Created initial cocoapods project.
