+++ .gitignore
@@ -1,2 +1,3 @@
-ImageRecognition.xcodeproj/xcuserdata/
+Example/ImageRecognition.xcworkspace/xcuserdata/
+Example/Pods/Pods.xcodeproj/xcuserdata/
