+++ Example/ImageRecognition.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
