+++ Example/ImageRecognition/ViewController.swift
@@ -1,0 +1,24 @@
+//
+//  ViewController.swift
+//  ImageRecognition
+//
+//  Created by Shahen on 10/16/2021.
+//  Copyright (c) 2021 Shahen. All rights reserved.
+//
+
+import UIKit
+
+class ViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        // Do any additional setup after loading the view, typically from a nib.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+}
+
