+++ Example/Podfile
@@ -1,0 +1,13 @@
+use_frameworks!
+
+platform :ios, '12.0'
+
+target 'ImageRecognition_Example' do
+  pod 'ImageRecognition', :path => '../'
+
+  target 'ImageRecognition_Tests' do
+    inherit! :search_paths
+
+    
+  end
+end
