--- ImageRecognition/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  ImageRecognition
-//
-//  Created by Arthur Poghosyan on 14.10.21.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
