+++ README.md
@@ -1,2 +1,29 @@
-# image-recognition-ios
-Image recognition library for IOS
+# ImageRecognition
+
+[![CI Status](https://img.shields.io/travis/stingle/ImageRecognition.svg?style=flat)](https://travis-ci.org/stingle/ImageRecognition)
+[![Version](https://img.shields.io/cocoapods/v/ImageRecognition.svg?style=flat)](https://cocoapods.org/pods/ImageRecognition)
+[![License](https://img.shields.io/cocoapods/l/ImageRecognition.svg?style=flat)](https://cocoapods.org/pods/ImageRecognition)
+[![Platform](https://img.shields.io/cocoapods/p/ImageRecognition.svg?style=flat)](https://cocoapods.org/pods/ImageRecognition)
+
+## Example
+
+To run the example project, clone the repo, and run `pod install` from the Example directory first.
+
+## Requirements
+
+## Installation
+
+ImageRecognition is available through [CocoaPods](https://cocoapods.org). To install
+it, simply add the following line to your Podfile:
+
+```ruby
+pod 'ImageRecognition'
+```
+
+## Author
+
+stingle, shahen@advancedtech.am
+
+## License
+
+ImageRecognition is available under the MIT license. See the LICENSE file for more info.
